requirements.txt
python-telegram-bot[job-queue]==21.6
psycopg[binary]==3.2.4
